+++ src/extension/routers/RegistrationAddAccountRouter/RegistrationAddAccountRouter.tsx
